+++ client/src/websocket/index.test.ts
@@ -21,7 +21,7 @@ import { describe, expect, it } from "vitest";
 
 import APIClient, { testClient as client } from "../api-client";
 import { StateModel, globalSchema } from "../model";
-import { NotificationsManager } from "../notifications";
+import NotificationsManager from "../notifications/NotificationsManager";
 import { sleep } from "../utils/helpers/HelperFunctions";
 import { WsServerMessage } from "./WsMessages";
 import {
